backend/pyproject.toml
[project]
name = "krisp-whisper-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "openai-whisper",
    "fastapi",
    "uvicorn[standard]",
    "python-multipart",
]
